Moved the AI Product creation banner to the products/ai package
